+++ CHANGELOG.md
@@ -5,6 +5,12 @@ see the commit logs. The format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Fish completion script.
+
 ## 0.3.1 - 2021-04-26
 
 ### Added